Delete system-specific build files
A system-specific build file is generated from the keil v5 project file, which has a file extension based on the name of the user for the system the project is built on.  This file is re-generated on each unique system when the project is built, so it is not needed in the repository.